requirements.txt
Django==4.2.9
django-crispy-forms==2.0
django-filter==23.3
djangorestframework==3.13.1
pywhatkit==5.3
reportlab==3.6.12
python-dateutil==2.8.2
Pillow==9.5.0
